% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DUMMY_DATA}
\alias{DUMMY_DATA}
\title{Dummy data used by AvE}
\format{
\subsection{\code{who}}{

A list of 1) A data.frame containing actual data, 2) An R object holding the
model that attempts to predict the actual data, 3) The field name that is
being predicted by the model
\describe{
\item{DATA}{The actual data with f}
\item{MODEL}{The model that predicts}
\item{RESPONSE}{The name of the field that is predicted by MODEL}
}
}
}
\source{
/data-raw/Prepare DUMMY_DATA.R
}
\usage{
DUMMY_DATA
}
\description{
Dummy data used by AvE
}
\keyword{datasets}
